* origin/master: (301 commits)
  KWSys Nightly Date Stamp
  Fix export logic so that vtkpng can build shared on CYGWIN.
  BUG: Correct the image stitching when boundaries are on.
  Disable FBO render passes with Mesa
  Fixed mismatch in BeginRenderProp and EndRenderProp for empty data.
  NetCDF cpp bindings have mixed long and size_t usage
  ENH: Fix regressions in the ATI-VolumeRendering-Bug commit
  KWSys Nightly Date Stamp
  ENH: Remove axis titles in this test.
  ENH: Allow mouse button action assigment.
  ENH: Introducing memset() to speedup initialization.
  BUG: Mixing up # of elements vs # of bytes
  STYLE: Clarify the units of _dataQuantity.
  Additional printself fix.
  ENH: Added ability to set mouse button functions.
  KWSys: Avoid passing string literal as char*
  COMP: Fixed shadowing warnings and indentation.
  BUG: Fixing uninitialized array.
  KWSys Nightly Date Stamp
  vtkHybrid: fix printself test failures.
  ...

+++ Utilities/vtknetcdf/netcdf.cpp
@@ -636,9 +636,9 @@ NcDim* NcVar::get_dim( int i ) const
     return the_file->get_dim(dims[i]);
 }
 
-long* NcVar::edges( void ) const  // edge lengths (dimension sizes)
+size_t* NcVar::edges( void ) const  // edge lengths (dimension sizes)
 {
-    long* evec = new long[num_dims()];
+    size_t* evec = new size_t[num_dims()];
     for(int i=0; i < num_dims(); i++)
       evec[i] = get_dim(i)->size();
     return evec;
@@ -806,7 +806,7 @@ NcValues* NcVar::get_rec(NcDim* rdim, long slice)
   return 0;
     }
 
-    long* edgel = edges();
+    size_t* edgel = edges();
     size_t* edge = new size_t[size];
     for (int i=1; i < size ; i++) {
   edge[i] = edgel[i];
@@ -899,7 +899,7 @@ NcBool NcVar::put_rec( NcDim* rdim, const TYPE* vals,                         \
     if (! result )                                                            \
       return FALSE;                                                           \
                                                                               \
-    long* edge = edges();                                                     \
+    size_t* edge = edges();                                                     \
     edge[idx] = 1;                                                            \
     result = put(vals, edge);                                                 \
     delete [] edge;                                                           \
@@ -921,7 +921,7 @@ long NcVar::rec_size(void) {
 long NcVar::rec_size(NcDim *rdim) {
     int idx = dim_to_index(rdim);
     long size = 1;
-    long* edge = edges();
+    size_t* edge = edges();
     for( int i = 0 ; i<num_dims() ; i++) {
   if (i != idx) {
     size *= edge[i];
@@ -1076,7 +1076,7 @@ NcVar_put_array(float)
 NcVar_put_array(double)
 
 #define NcVar_put_nd_array(TYPE)                \
-NcBool NcVar::put( const TYPE* vals, const long* count )          \
+NcBool NcVar::put( const TYPE* vals, const size_t* count )          \
 {                        \
     /* no need to check type() vs. TYPE, invoked C function will do that */   \
     if (! the_file->data_mode())                \
@@ -1089,7 +1089,7 @@ NcBool NcVar::put( const TYPE* vals, const long* count )          \
           ) == NC_NOERR;                                    \
 }
 
-NcBool NcVar::put( const ncbyte* vals, const long* count )
+NcBool NcVar::put( const ncbyte* vals, const size_t* count )
 {
     /* no need to check type() vs. TYPE, invoked C function will do that */
     if (! the_file->data_mode())
@@ -1102,7 +1102,7 @@ NcBool NcVar::put( const ncbyte* vals, const long* count )
           ) == NC_NOERR;
 }
 
-NcBool NcVar::put( const char* vals, const long* count )
+NcBool NcVar::put( const char* vals, const size_t* count )
 {
     /* no need to check type() vs. TYPE, invoked C function will do that */
     if (! the_file->data_mode())
@@ -1222,7 +1222,7 @@ NcVar_get_array(float)
 NcVar_get_array(double)
 
 #define NcVar_get_nd_array(TYPE)                \
-NcBool NcVar::get( TYPE* vals, const long* count ) const          \
+NcBool NcVar::get( TYPE* vals, const size_t* count ) const          \
 {                        \
     if (! the_file->data_mode())                \
       return FALSE;                    \
@@ -1234,7 +1234,7 @@ NcBool NcVar::get( TYPE* vals, const long* count ) const          \
           ) == NC_NOERR;     \
 }
 
-NcBool NcVar::get( ncbyte* vals, const long* count ) const
+NcBool NcVar::get( ncbyte* vals, const size_t* count ) const
 {
     if (! the_file->data_mode())
       return FALSE;
@@ -1244,7 +1244,7 @@ NcBool NcVar::get( ncbyte* vals, const long* count ) const
     return nc_get_vara_schar (the_file->id(), the_id, start,  (const size_t *) count, vals) == NC_NOERR;
 }
 
-NcBool NcVar::get( char* vals, const long* count ) const
+NcBool NcVar::get( char* vals, const size_t* count ) const
 {
     if (! the_file->data_mode())
       return FALSE;
